# Development/tools/machine_translation/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile '# Development/tools/machine_translation/requirements.in'
#
backoff==2.1.2
    # via -r # Development/tools/machine_translation/requirements.in
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   nltk
    #   sacremoses
colorama==0.4.4
    # via
    #   click
    #   tqdm
easynmt==2.0.2
    # via -r # Development/tools/machine_translation/requirements.in
fasttext==0.9.2
    # via easynmt
filelock==3.7.1
    # via
    #   huggingface-hub
    #   transformers
huggingface-hub==0.7.0
    # via transformers
idna==3.3
    # via requests
joblib==1.1.0
    # via
    #   nltk
    #   sacremoses
nltk==3.7
    # via easynmt
numpy==1.22.4
    # via
    #   easynmt
    #   fasttext
    #   transformers
packaging==21.3
    # via
    #   huggingface-hub
    #   transformers
protobuf==4.21.1
    # via easynmt
pybind11==2.9.2
    # via fasttext
pyparsing==3.0.9
    # via packaging
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
regex==2022.6.2
    # via
    #   nltk
    #   sacremoses
    #   transformers
requests==2.28.0
    # via
    #   huggingface-hub
    #   transformers
ruamel-yaml==0.17.21
    # via -r # Development/tools/machine_translation/requirements.in
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
sacremoses==0.0.53
    # via -r # Development/tools/machine_translation/requirements.in
sentencepiece==0.1.96
    # via easynmt
six==1.16.0
    # via sacremoses
tokenizers==0.12.1
    # via transformers
torch==1.11.0
    # via easynmt
tqdm==4.64.0
    # via
    #   -r # Development/tools/machine_translation/requirements.in
    #   easynmt
    #   huggingface-hub
    #   nltk
    #   sacremoses
    #   transformers
transformers==4.19.4
    # via easynmt
typing-extensions==4.2.0
    # via
    #   huggingface-hub
    #   torch
urllib3==1.26.9
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
